services/stock_comparison_agent/requirements.txt
fastapi==0.115.5
uvicorn==0.32.1
pydantic==2.10.3
alpaca-py==0.43.0
pandas==2.2.3
python-dotenv==1.0.1
